+++ NEWS.md
@@ -1,3 +1,6 @@
-# bellreg (development version)
+# bellreg 0.1.2
 
-* Added a `NEWS.md` file to track changes to the package.
+* The bellreg package now requires rstan version 2.26.
+* Implementation of extract_log_lik method for bellreg and zibellreg objects.
+* Implementation of new link functions (log, sqrt and identity for the count distribution and logit, probit, cloglog and cauchy for the degenerated distribution).
+* Implementation of AIC method for bellreg and zibellreg objects.